Refactor component name
