+++ src/components/pages/FanZonePage/FanZonePage.tsx
